requirements.txt
supabase
python-dotenv
requests
numpy
pandas
streamlit
plotly
scipy
scikit-learn
xgboost
lightgbm
